# docs/swagger.yaml
basePath: /
definitions:
...
  dto.TriggerParameter:
    properties:
      description:
        type: string
      example:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  dto.TriggerResponse:
    properties:
      category:
        type: string
      code:
        type: string
      description:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      parameters:
        items:
          $ref: '#/definitions/dto.TriggerParameter'
        type: array
    type: object
...
info:
  contact:
    email: support@ventros.com
    name: API Support
  description: API para gerenciamento de CRM com eventos e workflows
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ventros CRM API
  version: "1.0"
paths:
...
  /api/v1/automation/triggers:
    get:
      description: Retorna todos os triggers do sistema e customizados
      parameters:
      - description: Filtrar por categoria (session, message, pipeline, temporal,
          transaction, behavior)
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              custom_triggers:
                items:
                  $ref: '#/definitions/dto.TriggerResponse'
                type: array
              system_triggers:
                items:
                  $ref: '#/definitions/dto.TriggerResponse'
                type: array
            type: object
      summary: Lista triggers disponíveis
      tags:
      - Automation
  /api/v1/automation/triggers/{code}:
    get:
      description: Retorna metadados completos de um trigger incluindo parâmetros
        disponíveis
      parameters:
      - description: Código do trigger
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TriggerResponse'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Detalhes de um trigger
      tags:
      - Automation
  /api/v1/automation/triggers/custom:
    post:
      consumes:
      - application/json
      description: Permite que admins registrem triggers customizados com prefixo
        'custom.'
      parameters:
      - description: Metadados do trigger
        in: body
        name: trigger
        required: true
        schema:
          properties:
            code:
              type: string
            description:
              type: string
            name:
              type: string
            parameters:
              items:
                properties:
                  description:
                    type: string
                  name:
                    type: string
                  type:
                    type: string
                type: object
              type: array
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              message:
                type: string
              trigger:
                $ref: '#/definitions/dto.TriggerResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Registrar trigger customizado
      tags:
      - Automation
  /api/v1/automation/triggers/custom/{code}:
    delete:
      description: Remove trigger customizado previamente registrado
      parameters:
      - description: Código do trigger customizado
        in: path
        name: code
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Remover trigger customizado
      tags:
      - Automation
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key authentication for service-to-service communication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"